+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Prefix your message with one of the following:
 
 ## Unreleased
 
+- [Added] Parse configuration files as erb.
 - [Changed] `I18n.listen(run_on_start:)` was added to control if files should be
   exported during `I18n.listen`'s boot. The default value is `true`.